decoder: send data that received from underlying layer to render
Now, these generated code are commented out since we have not enabled
intergated environment yet.

TODO:
Do we need to send checksum - the problem is not whether to send, but
does the checksum behavior depends on some data of instance or render.

NOTE: this need to co-work with branch `increment_dev`.

Change-Id: I93f7440bba9de8a8abe8bd3ca6a0cec51cc4d3eb